demo_work/requirements.txt
Django~=4.0.4
ujson~=5.2.0
djangorestframework~=3.13.1
category~=0.1.0
requests~=2.27.1
PyYAML~=6.0
celery~=5.2.6
pytest~=7.1.2

django-allauth~=0.50.0
drf-spectacular~=0.22.1
drf-spectacular-sidecar~=2022.5.1
Pillow~=9.1.0
drf-compound-fields~=2.0.0

gunicorn
psycopg2-binary